+++ src/firebase/index.js
@@ -1,6 +1,23 @@
-import * as firebase from 'firebase/app';
-import 'firebase/database';
-import 'firebase/storage';
-import { db } from './config.js';
+// Import the functions you need from the SDKs you need
+import { initializeApp } from 'firebase/app';
+import { getFirestore } from 'firebase/firestore';
+// TODO: Add SDKs for Firebase products that you want to use
+// https://firebase.google.com/docs/web/setup#available-libraries
 
-export const database = db;
+// Your web app's Firebase configuration
+const firebaseConfig = {
+  apiKey: 'AIzaSyC5ckCyFMUdhFsdg-13Hg9phtEHVnBAgnM',
+  authDomain: 'draw-rush.firebaseapp.com',
+  databaseURL:
+    'https://draw-rush-default-rtdb.asia-southeast1.firebasedatabase.app',
+  projectId: 'draw-rush',
+  storageBucket: 'draw-rush.appspot.com',
+  messagingSenderId: '941454002806',
+  appId: '1:941454002806:web:89bae0d8f1eb7b7fb76d57',
+};
+
+// Initialize Firebase
+const app = initializeApp(firebaseConfig);
+const db = getFirestore(app);
+
+export default db;
