+++ src/index.js
@@ -5,9 +5,9 @@ import App from './App';
 import reportWebVitals from './reportWebVitals';
 import { Provider, ReactReduxContext } from 'react-redux';
 import './style.css';
-import AppChat from './components/AppChat';
+import Chat from './components/chat';
 
-ReactDOM.render(<AppChat />, document.getElementById('root'));
+ReactDOM.render(<Chat />, document.getElementById('root'));
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
